Identifier cartouche (#4959)
* pass in attributeExpression to the component controls

* Update component-section.tsx

* oops

* IdentifierExpressionCartoucheControl

* data picker opens on clicking the cartouche

* Update property-control-controls.tsx

* moving test jsIdentifier to useComponentPropsInspectorInfo

* adding originalJavascript to jsIdentifier and co

* rendering more kinds of expressions with the cartouche

* fancy leading ellipsis

* Revert "adding originalJavascript to jsIdentifier and co"

This reverts commit e6a0359db2a298f428c3900ee3b7003e4f2d09b3.

* hey it is actually working now?!

* removing the X for now

* remove placeholder  hack

* for now, only show cartouche on base level

* updating perf snapshots

* moving jsxElementChildToText to dedicated file

* do not override the jsx control for inline jsx code

* show expressions as 'Expression'

* Update editor/src/components/inspector/common/property-controls-hooks.ts


---------
