+++ editor/src/core/performance/performance-regression-tests.spec.tsx
@@ -127,7 +127,7 @@ describe('React Render Count Tests -', () => {
 
     const renderCountAfter = renderResult.getNumberOfRenders()
     // if this breaks, GREAT NEWS but update the test please :)
-    expect(renderCountAfter - renderCountBefore).toMatchInlineSnapshot(`855`)
+    expect(renderCountAfter - renderCountBefore).toMatchInlineSnapshot(`845`)
     expect(renderResult.getRenderInfo()).toMatchSnapshot()
   })
 
@@ -249,7 +249,7 @@ describe('React Render Count Tests -', () => {
 
     const renderCountAfter = renderResult.getNumberOfRenders()
     // if this breaks, GREAT NEWS but update the test please :)
-    expect(renderCountAfter - renderCountBefore).toMatchInlineSnapshot(`623`)
+    expect(renderCountAfter - renderCountBefore).toMatchInlineSnapshot(`624`)
     expect(renderResult.getRenderInfo()).toMatchSnapshot()
   })
 })
